Heartbeat and known server list sync

+++ ConsistentReplicas/src/org/umn/distributed/consistent/common/Props.java
@@ -19,22 +19,22 @@ public class Props {
 	public static int COORDINATOR_PORT;
 	public static int COORDINATOR_SERVER_THREADS;
 
-	public static int HEARTBEAT_INTERVAL;
+	public static long HEARTBEAT_INTERVAL;
 	public static int REMOVE_INTERVAL;
 
 	public static int NETWORK_TIMEOUT;
-	
+
 	public static int QUORUM_SYNC_TIME_MILLIS = 5000;
 
-//	static{
-//		loadProperties("config.properties");
-//	}
-//	
+	// static{
+	// loadProperties("config.properties");
+	// }
+	//
 	public static void loadProperties(String propertyFile) {
 		Properties prop = new Properties();
 		try {
 			prop.load(new FileInputStream(propertyFile));
-			ENCODING = prop.getProperty("encoding","UTF8");
+			ENCODING = prop.getProperty("encoding", "UTF8");
 
 			// TODO: validate the port numbers before using
 
@@ -58,14 +58,14 @@ public class Props {
 			COORDINATOR_PORT = Integer.parseInt(prop
 					.getProperty("coordinatorPort"));
 
-			HEARTBEAT_INTERVAL = Integer.parseInt(prop
+			HEARTBEAT_INTERVAL = Long.parseLong(prop
 					.getProperty("heartbeatInterval"));
 			REMOVE_INTERVAL = Integer.parseInt(prop
 					.getProperty("deregisterInterval"));
 			NETWORK_TIMEOUT = Integer.parseInt(prop
 					.getProperty("totalNetworkTimeout"));
-			QUORUM_SYNC_TIME_MILLIS = Integer.parseInt(prop
-					.getProperty("quorumSyncTimeout", "10000"));
+			QUORUM_SYNC_TIME_MILLIS = Integer.parseInt(prop.getProperty(
+					"quorumSyncTimeout", "10000"));
 		} catch (IOException ex) {
 			ex.printStackTrace();
 		}
